refactor: move assets into their own package
This refactoring will split nodeup configuration out of cloudup, to
enable us to generate a nodeup configuration more easily, for example
in kops-controller.  That in turn will allow more dynamic usage with
e.g. cluster-api
